MCP_backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Database
psycopg2-binary==2.9.9

# LLM Providers
openai>=1.40.0
anthropic>=0.34.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
comtypes>=1.2.0

# CORS
python-jose[cryptography]==3.3.0
